requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
sqlalchemy==2.0.36
alembic==1.13.3

# Web scraping
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0

# Task queue
celery==5.4.0
redis==5.1.1

# Telegram bot
python-telegram-bot==21.7

# Google Cloud / Sheets
google-cloud-storage==2.18.2
google-auth==2.35.0
google-api-python-client==2.149.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0

# Image processing
Pillow==11.0.0
opencv-python-headless==4.10.0.84

# Dashboard
streamlit==1.39.0
pandas==2.2.3

# Utilities
python-dotenv==1.0.1
httpx==0.27.2
tenacity==9.0.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-mock==3.14.0
pytest-cov==6.0.0
